Fixed #15
Added start screen that allows user to display scoreboard, start game, or exit. Have not implemented MVVM structure yet, but have already added folder for it. Added field in CollisionMan to keep track of all timers, so that in code behind when gtame is over or won, all the timers stop. Also added HighScoreDialog to display a content dialog with a listview.